+++ ChangeLog.md
@@ -3,6 +3,10 @@
 
 2021/08/27
 
+1. Change sidenav visibilty rulue
+
+2021/08/27
+
 1. Fix bug that sidenav didn't hide after redirecting and renameing some variable
 2. Implement a simple 404 page, and redirection